refactor: use events for lose coupling between doctor service and appointments service

+++ src/services/doctors-service.ts
@@ -6,12 +6,16 @@ import CreateDoctorDto from "../dto/doctors/create-doctor-dto";
 import { plainToClass } from "class-transformer";
 import UpdateDoctorDto from "../dto/doctors/update-doctor-dto";
 import { merge } from "lodash";
+import { ServiceEventEmitter } from "./service-event-emitter";
+import { ServiceEvent } from "../enums/service-event";
 
 export default class DoctorsService {    
-    constructor(
-        private readonly repository: DoctorsRepository
-    ) {
-        
+    private readonly repository: DoctorsRepository;
+    private readonly eventEmitter: ServiceEventEmitter;
+
+    constructor(repository: DoctorsRepository, eventEmitter: ServiceEventEmitter) {
+        this.repository = repository;
+        this.eventEmitter = eventEmitter;
     }
 
     public async createDoctor(doctorDto: CreateDoctorDto): Promise<DoctorModel> {
@@ -44,9 +48,16 @@ export default class DoctorsService {
         return this.repository.update(doctor);
     }
 
-    public async deleteById(id: string): Promise<DoctorModel | undefined> {
+    public async deleteDoctorById(id: string): Promise<DoctorModel | undefined> {
         const doctor = await this.repository.get(id);
-        return await this.repository.remove(doctor);
+        if (!doctor) {
+            return;
+        }
+        const deleted = await this.repository.remove(doctor);
+        if (deleted) {
+            this.eventEmitter.emit(ServiceEvent.DOCTOR_DELETED, id);
+        }
+        return deleted;
     }
 }
